Update installation.md curl call for PowerShell
curl is included with Windows now but curl is also an alias for for the cmdlet Invoke-WebRequest in Windows PowerShell (not PowerShell Core).
Because of this, calling bare curl will use the alias first. By appending .exe to curl to get curl.exe, the binary expected will be called instead and the proper outcome happens.